+++ core/src/test/java/com/zextras/carbonio/files/api/notifications/RemovedNodeNotificationTrashApiIT.java
@@ -16,7 +16,7 @@ import com.zextras.carbonio.files.dal.dao.ebean.ACL;
 import com.zextras.carbonio.files.dal.repositories.interfaces.FileVersionRepository;
 import com.zextras.carbonio.files.dal.repositories.interfaces.LinkRepository;
 import com.zextras.carbonio.files.dal.repositories.interfaces.NodeRepository;
-import com.zextras.carbonio.files.utilities.TestFilesConfig;
+import com.zextras.carbonio.files.utilities.MockFilesConfig;
 import com.zextras.carbonio.files.utilities.http.HttpRequest;
 import com.zextras.carbonio.files.utilities.http.HttpResponse;
 import org.assertj.core.api.Assertions;
@@ -156,7 +156,7 @@ class RemovedNodeNotificationTrashApiIT {
   @Test
   void givenANodeRemovalByTrashingItOnASharedDirectoryAndDisabledNotificationsNoNotificationShouldBeSavedOrReturned() {
     // Given
-    ((TestFilesConfig)
+    ((MockFilesConfig)
         simulator
             .getInjector()
             .getInstance(FilesConfig.class))
@@ -187,7 +187,7 @@ class RemovedNodeNotificationTrashApiIT {
     Assertions.assertThat(notifications).hasSize(0);
 
     //reset
-    ((TestFilesConfig)
+    ((MockFilesConfig)
         simulator
             .getInjector()
             .getInstance(FilesConfig.class))
