+++ core/src/test/java/com/zextras/carbonio/files/dal/repositories/impl/ebean/NodeRepositoryEbeanIT.java
@@ -5,7 +5,7 @@
 package com.zextras.carbonio.files.dal.repositories.impl.ebean;
 
 import com.zextras.carbonio.files.dal.FilesPostgreSQLContainer;
-import com.zextras.carbonio.files.utilities.TestFilesConfig;
+import com.zextras.carbonio.files.utilities.MockFilesConfig;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
@@ -22,7 +22,7 @@ class NodeRepositoryEbeanIT {
 
   @BeforeAll
   static void setup() {
-    database = new FilesPostgreSQLContainer(new TestFilesConfig());
+    database = new FilesPostgreSQLContainer(new MockFilesConfig());
   }
 
   @BeforeEach
