+++ core/src/test/java/com/zextras/carbonio/files/rest/repositories/UserRepositoryRestTest.java
@@ -6,7 +6,7 @@ package com.zextras.carbonio.files.rest.repositories;
 
 import com.zextras.carbonio.files.cache.CacheHandler;
 import com.zextras.carbonio.files.dal.repositories.impl.ebean.UserRepositoryRest;
-import com.zextras.carbonio.files.utilities.TestFilesConfig;
+import com.zextras.carbonio.files.utilities.MockFilesConfig;
 import com.zextras.carbonio.usermanagement.UserManagementClient;
 import com.zextras.carbonio.usermanagement.entities.UserId;
 import com.zextras.carbonio.usermanagement.entities.UserMyself;
@@ -28,7 +28,7 @@ class UserRepositoryRestTest {
 
   @BeforeEach
   void setup() {
-    userRepositoryRest = new UserRepositoryRest(new TestFilesConfig(), Mockito.mock(CacheHandler.class));
+    userRepositoryRest = new UserRepositoryRest(new MockFilesConfig(), Mockito.mock(CacheHandler.class));
   }
 
   @Test
